requirements.txt
tkinter
Pillow
pyinstaller==5.1